pyproject.toml
[project]
name = "sae-chat"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "eai-sparsify",
    "hf-xet>=1.1.5",
    "langchain-text-splitters>=0.3.9",
    "pandas>=2.3.1",
    "pymilvus>=2.5.14",
    "sentence-transformers>=5.0.0",
    "torch>=2.7.1",
    "tqdm>=4.67.1",
]

[tool.uv.sources]
eai-sparsify = { git = "https://github.com/EleutherAI/sparsify" }

[tool.uv.workspace]
members = ["sae-chat"]
